Convert dynamic network's capacity to longs. FluidNetwork still clamped to integer for now though. Also fix some potentially incorrect calculations for energy network scale at large network capacities

+++ src/main/java/mekanism/common/capabilities/basic/DefaultLogisticalTransporter.java
@@ -119,7 +119,7 @@ public class DefaultLogisticalTransporter implements ILogisticalTransporter {
     }
 
     @Override
-    public int getTransmitterNetworkCapacity() {
+    public long getTransmitterNetworkCapacity() {
         return 0;
     }
 
@@ -130,7 +130,7 @@ public class DefaultLogisticalTransporter implements ILogisticalTransporter {
     }
 
     @Override
-    public int getCapacity() {
+    public long getCapacity() {
         return 0;
     }
